tests: Create positive Mesh shader file
